Generics.Deriving.ConNames

Description

Functionality for retrieving the names of the possible contructors

class ConNames f where

Methods

gconNames :: f a -> [String]

gconNameOf :: f a -> String

conNames :: (Generic a, ConNames (Rep a)) => a -> [String]

Return the name of all the constructors of the type of the given term.

conNameOf :: (ConNames (Rep a), Generic a) => a -> String

Return the name of the constructor of the given term