Generics.Deriving.Monoid

Description

This module provides two main features:

GMonoid, a generic version of the Monoid type class, including instances of the types from Data.Monoid

Default generic definitions for the Monoid methods mempty and mappend

The generic defaults only work for types without alternatives (i.e. they have only one constructor). We cannot in general know how to deal with different constructors.

GMonoid type class

class GMonoid a where

Minimal complete definition

Nothing

Methods

gmempty :: a

Generic mempty

gmappend :: a -> a -> a

Generic mappend

gmconcat :: [a] -> a

Generic mconcat

gmempty :: (Generic a, GMonoid' (Rep a)) => a

Generic mempty

gmappend :: (Generic a, GMonoid' (Rep a)) => a -> a -> a

Generic mappend

Default definitions

GMonoid

gmemptydefault :: (Generic a, GMonoid' (Rep a)) => a

gmappenddefault :: (Generic a, GMonoid' (Rep a)) => a -> a -> a

Internal auxiliary class for GMonoid

class GMonoid' f where

Methods

gmempty' :: f x

gmappend' :: f x -> f x -> f x

Monoid

These functions can be used in a Monoid instance. For example:

 -- LANGUAGE DeriveGeneric

 import Generics.Deriving.Base (Generic)
 import Generics.Deriving.Monoid

 data T a = C a (Maybe a) deriving Generic

 instance Monoid a => Monoid (T a) where
   mempty  = memptydefault
   mappend = mappenddefault
 

memptydefault :: (Generic a, Monoid' (Rep a)) => a

mappenddefault :: (Generic a, Monoid' (Rep a)) => a -> a -> a

Internal auxiliary class for Monoid

class Monoid' f where

Methods

mempty' :: f x

mappend' :: f x -> f x -> f x

The Monoid module

This is exported for convenient access to the various wrapper types.

module Data.Monoid