Generics.Deriving.Enum

Generic enum class

class GEnum a where

Methods

genum :: [a]

genum :: (Generic a, Enum' (Rep a)) => [a]

Default definitions for GEnum

genumDefault :: (Generic a, Enum' (Rep a)) => [a]

toEnumDefault :: (Generic a, Enum' (Rep a)) => Int -> a

fromEnumDefault :: (GEq a, Generic a, Enum' (Rep a)) => a -> Int

Internal enum class

class Enum' f where

Minimal complete definition

enum'

Methods

enum' :: [f a]

Generic Ix class

class Ord a => GIx a where

Methods

range :: (a, a) -> [a]

The list of values in the subrange defined by a bounding pair.

index :: (a, a) -> a -> Int

The position of a subscript in the subrange.

inRange :: (a, a) -> a -> Bool

Returns True the given subscript lies in the range defined the bounding pair.

range :: (GEq a, Generic a, Enum' (Rep a)) => (a, a) -> [a]

The list of values in the subrange defined by a bounding pair.

index :: (GEq a, Generic a, Enum' (Rep a)) => (a, a) -> a -> Int

The position of a subscript in the subrange.

inRange :: (GEq a, Generic a, Enum' (Rep a)) => (a, a) -> a -> Bool

Returns True the given subscript lies in the range defined the bounding pair.

Default definitions for GIx

rangeDefault :: (GEq a, Generic a, Enum' (Rep a)) => (a, a) -> [a]

indexDefault :: (GEq a, Generic a, Enum' (Rep a)) => (a, a) -> a -> Int

inRangeDefault :: (GEq a, Generic a, Enum' (Rep a)) => (a, a) -> a -> Bool